Removed ValidableObject class
